use rtgl custom path
