+++ package.json
@@ -16,14 +16,12 @@
     "vite": "^6.3.5"
   },
   "scripts": {
-    "setup:web": "node scripts/build-setup.js web",
-    "setup:tauri": "node scripts/build-setup.js tauri",
-    "build:dev": "bun run setup:web && rm -rf _site && mkdir _site && cp -rf static/* _site/ && rtgl ui build-svg && rtgl fe build -d",
-    "build": "bun run setup:web && rm -rf _site && mkdir _site && cp -rf static/* _site/ && rtgl ui build-svg && rtgl fe build",
-    "build:web": "bun run setup:web && rm -rf _site && mkdir _site && cp -rf static/* _site/ && rtgl ui build-svg && rtgl fe build",
-    "build:tauri": "bun run setup:tauri && rm -rf _site && mkdir _site && cp -rf static/* _site/ && rtgl ui build-svg && rtgl fe build",
-    "watch": "bun run setup:web && rtgl fe watch",
-    "watch:tauri": "bun run setup:tauri && rtgl fe watch",
+    "build:dev": "rm -rf _site && mkdir _site && cp -rf static/* _site/ && rtgl ui build-svg && rtgl fe build -d -s src/setup.web.js",
+    "build": "rm -rf _site && mkdir _site && cp -rf static/* _site/ && rtgl ui build-svg && rtgl fe build -s src/setup.web.js",
+    "build:web": "rm -rf _site && mkdir _site && cp -rf static/* _site/ && rtgl ui build-svg && rtgl fe build -s src/setup.web.js",
+    "build:tauri": "rm -rf _site && mkdir _site && cp -rf static/* _site/ && rtgl ui build-svg && rtgl fe build -s src/setup.tauri.js",
+    "watch": "rtgl fe watch -s src/setup.web.js",
+    "watch:tauri": "rtgl fe watch -s src/setup.tauri.js",
     "serve": "bunx serve _site -p 3001",
     "lint": "bunx prettier --check src",
     "lint:fix": "bunx prettier --check --write src",
